+++ include/private/kmMqtt/Utils/Deferrer.h
@@ -8,6 +8,7 @@
 
 #include "kmMqtt/Utils/Event.h"
 #include "kmMqtt/GlobalMacros.h"
+#include "kmMqtt/STL/KmMemory.h"
 #include <queue>
 #include <mutex>
 #include <type_traits>
@@ -55,7 +56,7 @@ namespace kmMqtt
 			template<typename TFunc>
 			void defer(TFunc&& event)
 			{
-				m_events.emplace(std::make_unique<Callable<TFunc>>(std::forward<TFunc>(event)));
+				m_events.emplace(kmStd::make_unique<Callable<TFunc>>(std::forward<TFunc>(event)));
 			}
 
 			void invokeEvents() noexcept;
@@ -67,7 +68,7 @@ namespace kmMqtt
 			}
 
 		private:
-			std::queue<std::unique_ptr<ICallable>> m_events;
+			std::queue<kmStd::unique_ptr<ICallable>> m_events;
 		};
 	}
 }