+++ src/MqttClient.cpp
@@ -15,7 +15,7 @@ namespace kmMqtt
 			auto envFactory{ DefaultEnvironmentFactory() };
 			auto env{ envFactory.createEnvironment() };
 			const MqttClientOptions defaultOptions{};
-			m_impl = std::make_unique<MqttClientImpl>(env, defaultOptions);
+			m_impl = kmStd::make_unique<MqttClientImpl>(env, defaultOptions);
 
 			if (defaultOptions.getTickMode() == TickMode::ASYNC)
 			{
@@ -26,7 +26,7 @@ namespace kmMqtt
 		}
 
 		MqttClient::MqttClient(const IMqttEnvironment* const env, const MqttClientOptions& clientOptions)
-			: m_impl(std::make_unique<MqttClientImpl>(env, clientOptions))
+			: m_impl(kmStd::make_unique<MqttClientImpl>(env, clientOptions))
 		{
 			if (clientOptions.getTickMode() == TickMode::ASYNC)
 			{